New version of TTreeFormula by Philippe Canal. This version should be properly handling the following cases. Draw("fVertex"); Draw("fVertex[10]"); Draw("fVertex[][2]"); Draw("fVertex[600]"); Draw("fVertex[590][1]");
It should also handle properly operation including arrays of the same
number of dimension but different size.

E.g. with

carre[2][3] and cube[3][4]

Draw("carre - cube");

should be doing the equivalent of

for(i=0;i<2;i++)
  for(j=0;j<3;j++)
        histogram( carre[i][j] - cube[i][j] );

It should be the case both is the first dimension is a real dimension
of one 'acquired' from being in a TClonesArray.


git-svn-id: http://root.cern.ch/svn/root/trunk@200 27541ba8-7e3a-0410-8455-c3a389f83636