Fix a possible memory leak in TF1::Copy (thanks Martin.Woudstra). It creates new arrays for the parameters, but does not delete the old ones.
This is fine if the TF1 to be copied into is a newly
created one with the default constructor.
But if not, this looks like a memory leak...

I know the comment says: "Copy this F1 to a new F1",
but nobody can guarantee that this function is called
this way! And 'new' is not the same as 'empty'.


git-svn-id: http://root.cern.ch/svn/root/trunk@1089 27541ba8-7e3a-0410-8455-c3a389f83636