import of CINT 5.14.63 For what is new see http://root.cern.ch/root/Cint.phtml?relnote.
This version might generate a warning in several dictionaries, like:
hist/src/G__Hist.cxx:1724: warning: `void G__operator_delete(void *)' defined but not used.
This will be fixed in the next version, but it is imported now since we need
testing of this new version of CINT on all platforms. If things don't compile
undefine G__EXPERIMENTAL1423 in G__ci.h and let us know.


git-svn-id: http://root.cern.ch/svn/root/trunk@1098 27541ba8-7e3a-0410-8455-c3a389f83636