New version of TTreeFormula/TFormula by Philippe Canal: In a TFormula . and -> are now synonymous.
In a TTreeFormula, you can now specify the index of an array (TClonesArray
or regular array) via another expression of the variable of the tree.

For example the following are now correctly handled:
  myTree->Draw("clone1.a1:clone2.a2[2]");
  myTree->Draw("clone1.a1:clone2.a2[clone1.b1]");
  myTree->Draw("clone1.a1","clone2.a2[clone1.b1]>0");

This patch also enable (rudimentarily for now) the access to data members.
So now both the following are possible:
  T->Draw("event->GetNtrack()");
  T->Draw("event.fNtrack");

The lookup of leafname has been improved so that whether of not the dot
has been forced into the branchname the following will work:

  T->Draw("event.fTemperature")
  T->Draw("fTemperature")


git-svn-id: http://root.cern.ch/svn/root/trunk@1569 27541ba8-7e3a-0410-8455-c3a389f83636