From Philippe: This version of TTreeFormula enables using method calls within the Draw command with new style branches.
Together with the patch for TStreamerInfo, this allow stress to
work properly in the new style of branches.

However it does not completely fix the access to data member.
It is currently limited to public data member.  We need to
find a function that return the offset of a datamember given
its name (this function needs to lookup in the class itself
but also in its base).


git-svn-id: http://root.cern.ch/svn/root/trunk@1993 27541ba8-7e3a-0410-8455-c3a389f83636