several enhancements from Philippe. TFormula:  added support for the shift operator '>>' and '<<'. However the associativeness is currently wrong.  It should have been left associative (i.e.  b << 4 >> 8 == (b<<4)>>8 ).
TTreePlayer: add support for letting '>>' be interpreted as a shift operator.
Now only the last '>>' which is NOT is parenthesis or brackets is interpreted
as meaning 'feed an histogram'.  This might(?) be improved if there is a way
to guess is the right hand parameter of '>>' is and histogram or a variable.

TTreePlayer: enable string comparaison when the new style of branches and
with data member reading.  (Are left to do: method result).


git-svn-id: http://root.cern.ch/svn/root/trunk@2382 27541ba8-7e3a-0410-8455-c3a389f83636