The following mods by Mathieu de Naurois. The following was an email from Matthieu
"You set the Thread Id at two different places.
1) in Int_t TPosixThread::Run(TThread *th)

  th->fId = id; SetUniqueID(th->fId);

2)in void *TThread::Fun(void *ptr)

  th = (TThread *)ptr;
  th->fId = SelfId(); th->SetUniqueID(th->fId);

The second solution is the only good one, because the Thread uses its it
in the TThread::Fun
function. So if TPosixThread::Run was not fast enough, you get a clash.

BUT, in TThread::Fun(void *ptr) you use the Id before assigning it, in
the following call
  CleanUpPush((void *)&AfterCancel,th);  // Enable standard cancelling
function

So, I suggest to move the Id definition before that call."


git-svn-id: http://root.cern.ch/svn/root/trunk@2385 27541ba8-7e3a-0410-8455-c3a389f83636