Update by Philippe of TEventList and TTreePlayer such that one can reproduce the results obtained at the creation of the list.
The Title of the EventList object is now fully used to store the
actual selection criteria used to produce the EventList.  When
SetReapplyCut is set to true, the criteria is fed back into the
new selection criteria.

This allows for example to really make a selection of some of
the tracks of the events and continue analysis on just them.

For example:

  T->Draw(">>pxlist","fPx>2.7");

creates a TEventList 'pxlist' that list all events that have one or
more tracks with fPx greater than 2.7

  T->SetEventList(pxlist);
  T->Draw("fPx");

will draw the fPx of ALL the tracks of the events that have at least
one track with a fPx greater than 2.7

  pxlist->SetReapplyCut(kTRUE);
  T->SetEventList(pxlist);
  T->Draw("fPx");

will draw the fPx of ONLY the tracks that ha a fPx greater than 2.7.

Also

  T->Draw(">>+pxlist","fPy>2.7");

Will have pxlist.fTitle == (fPx>2.7)||(fPx>2.5)

Whenever, a TEventList is filled, its title is properly maintained
to reflect exactly the selection(s) used to create it.

Note that as a side-effect , the TEventList will now contain each entry
only once.  (it was possible to add an entry multiple times in the list).

CCVS: ----------------------------------------------------------------------


git-svn-id: http://root.cern.ch/svn/root/trunk@3083 27541ba8-7e3a-0410-8455-c3a389f83636