Add new static global member fgRejectPoint with the new static functions:   void TF1::RejectPoint(Bool_t reject=kTRUE)   Bool_t TF1::RejectedPoint()
The function TF1::RejectPoint may be used in a user fitting function
to reject points. The fgRejectPoint is automatically reset by Root.

A new member function TF1::IsInside has been added.
Currently it returns kTRUE if the point is inside the function limits.
This function is redefined in TF2 and TF3.
The main purpose of this function is to give additional possibilities
in functions deriving from TF1. For example, if a function is only
defined in some specific non-contiguous ranges, this could be
implemented in a class deriving from TF1 and simply redefining
the function TF1::IsInside.


git-svn-id: http://root.cern.ch/svn/root/trunk@3120 27541ba8-7e3a-0410-8455-c3a389f83636