Christion Holm Christensen has made the following improvements: I've just realised that I've done something extreemly stupid in TPrincipal::MakeHistograms which can make one waste enourmous amounts of space and time.  The problem was, that the method looped over the user data wether or not you actually wanted to plot it or not, thereby wasting a lot of time.  Also, I allocated arrays via new in each iteration, which is ofcourse a waste of time, so I moved that outside the loop and added a zeroing loop instead.
Further,  I added the option 'd' to the CTOR, so that one may opt to
_not_ store the input data.  This reguired a new data member, and
hence a version bumb.  The default is to store the input data.  I
think this is a useful option for some, since the data arrays can
become quite large.

Thirdly, I modified the histogram names to be 0 padded upto 3 digits.
That is, the format is "%03d" rather than "%d", since that allows for
better overview in the browser.


git-svn-id: http://root.cern.ch/svn/root/trunk@3326 27541ba8-7e3a-0410-8455-c3a389f83636