Delete function FindGoodLimits. Use services of THLimitsFinder in TTreePlayer::SetEstimate.
Modify DrawSelect to use the modified function TTreeFormula::SetAxis.
This new function takes care of setting the bit kIsInteger in the axis.
This also fixes a problem when histograming integer variables. The number of bins
was not correctly recomputed.

Add a fix from Philippe in MakeClass to initialize members holding object pointers.


git-svn-id: http://root.cern.ch/svn/root/trunk@3669 27541ba8-7e3a-0410-8455-c3a389f83636