New feature added by Philippe: I am taking advantage of the fact that we have one painter per histogram to individualize a little better the stats box.
The HistPainter now handles a pointer to the stats box.  In particular this
means that you can rename it and still have it being updated.  The stats can
also now be retrieved via hist->GetPainter()->GetStats();  The HistPainter
now keeps contact with the PaveStats box whether its name have changed or not.


git-svn-id: http://root.cern.ch/svn/root/trunk@3915 27541ba8-7e3a-0410-8455-c3a389f83636