From Olivier: - Bug fixed in PaintPalette. Hparam.zmin and Hparam.zmax were used instead of the real minimum and maximum of the histogram. This produced wrong palette painting. The zmin and zmax in Hparam were used to handle Log option on z axis. This is not necessary because if the maximum of the histogram is <= 0 nothing is painted, and in the case where maximum is >0 and the minimum <= 0, TH1::GetMinimum returns a positive value equal to max/1000.
- 2d functions painted with option COL produced empty boxes instead of
filled color boxes. This was a side effect of a change in the function
constructor in which the fill style is now set to "empty".
fH->SetFillStyle(1);
has been added in PaintColorLevels to fix that.

- In DefineColorLevels "theColor" should be computed as:
theColor = Int_t((i+0.99)....
to be compatible with the same value computed in PaintPalette

- A similar problem occurred in PaintContour. "theColor" was computed as:
Int_t((ipoly+1.99)... instead of: Int_t((ipoly+0.99) ...


git-svn-id: http://root.cern.ch/svn/root/trunk@4821 27541ba8-7e3a-0410-8455-c3a389f83636