This patch to ROOT and CINT allows the user to use ClassDef with either the current syntax:         ClassDef(Class,Version) // Comment or add a 'semi-colon':         ClassDef(Class,Version); // Comment
This is very usefull for text editor to properly automatically ident,
color, etc. the classes.

I also changed CINT's behavior regarding ClassDef when interpreting the
ClassDef macro (in the ROOT context).  Currently, if a macro which is
interpreted contains the ClassDef macro, it fails.  I added some support
so that the ClassDef macro is ignore (since anyway in interpreted mode
it is currently more or so pointless). By Philippe.


git-svn-id: http://root.cern.ch/svn/root/trunk@5017 27541ba8-7e3a-0410-8455-c3a389f83636