From Philippe: "I think it is inconvenient for the user (and the documenter) that the 'cint' executable behaves differently in the ROOT environment or in the pure CINT environment.  For example, the lastest change broke Fons' script to build the cintdlls (since linG__set_globalcompk was off by default, the cintdlls where actually empty!).
So I propose to reset the 'link mode' to be on by default for the 'cint'
executable.  We can then explicitly set it to be off in the 'rootcint'
executable.

In order to properly implement this, I also needed to fix the handling
of the argument to '-c' (corretly use the abs value and propagate all
the change into G__set_globalcomp.


git-svn-id: http://root.cern.ch/svn/root/trunk@5465 27541ba8-7e3a-0410-8455-c3a389f83636