Applied rootd patch by Ulrik Egede and Tim Adye from BaBar. New features:
1) New option, -P FILE, to specify an alternate SRP password file. As an
   example

   rootd -P $HOME/.srootdpass2

   will start rootd using the files $HOME/.srootdpass2.conf and
   $HOME/.srootdpass2 for SRP authentication.

2) Modified option, -p port1-port2, will search for the first available
port in the range. You can also specify -p 0-N for search relative to the
service port specified in /etc/services (this was sort of accidental, but
not a bad feature :-). If a single port is specified as before, then no
search is made.

3) Unless started by inetd (rootd -i), it prints something like

    ROOTD_PORT=5151
    ROOTD_PID=14433

before spawning the daemon so the user knows what was used (eval `rootd`
will set these as variables in Bourne-shells).

4) rootd now shows an error message (as well as the syslog message it
always sent) if there is any problem binding the port or forking the
daemon.


git-svn-id: http://root.cern.ch/svn/root/trunk@5507 27541ba8-7e3a-0410-8455-c3a389f83636