m andrei; TGeoManager :  - added TGeoManager::Safety() that computes the global safety for the current point. This is just the first implementation, since it does not use yet voxelization. This will be needed for the VMC because for the time being safety is computed just for the crossed volumes along the current direction plus the current volume, which of course is not enough - finding the closest neighbour for a given point has to use a different algorithm.  - added weight estimation for the top volume
TGeoMaterial :
 - material index in the list of materials added as data member (as for
volumes) - I needed this for weight computation, but it will be helpfull
for any other analysis based on material indexing.

TGeoVolume :
 - weight estimation method in the context menu of volumes : can be
verbose (default) or not and takes as argument the desired precision.
Prints error (1 x SIGMA)

TGeoChecker:
 - implementation of weight estimation based on sampling random points
in the bounding box of a volume. Materials with rho<0.01 g/cm3 (mostly
gases) ignored due to the fact that we generally do not want to weight
the air in the top volume container (it can have quite significant mass)

other modified classes just provide redirection of Weight() method to
the geopainter package.


git-svn-id: http://root.cern.ch/svn/root/trunk@6136 27541ba8-7e3a-0410-8455-c3a389f83636