From Philippe and Brett Wiren: The problem is that when writing the line with G__setnewtype, cint/rootcint tries to grab a comment for the typedef (so far so good).  This is done by calling G__getcommenttypedef which read up to G__ONELINE-1 and stores it in its parameters which in this case is of size G__ONELINE (so, so far still no problem).  However, it is then massaged by passing the same buffer to G__add_quotation which adds at least 2 characters.
In addition in your case, the typedef is a function typedef without any explicit comment.  In this case, cint/rootcint comments the typedef by its declarations.  However it starts from the beginning of the typedef to the end of the line with the typedef.  Since your typedef is inside a macro, this line is longer than G__ONELINE.

Hence the slight out-of-bounds write.

Attached is a patch that insures that a typedef comments terminates at the first ';'.  This solves your problems, however, in other cases this might chop the user specified comment if it happens to contains a ';' ....


git-svn-id: http://root.cern.ch/svn/root/trunk@6355 27541ba8-7e3a-0410-8455-c3a389f83636