From Philippe: In TBranchElement::SetAddress we have something like:
     TClass *clm = gROOT->GetClass(GetClassName());
     Int_t *offsets = clm->GetStreamerInfo()->GetOffsets();
     ...
        if (mOffset > 0) fOffset = mOffset -offsets[fID];

However, if the class has been modified so that the order of the data member in the "live" class is different from
the version that was used to write the file, the fID does NOT point to the index of the needed datamember in the new
StreamerInfo.  We have to retrieve it by name instead (hence the patch).


git-svn-id: http://root.cern.ch/svn/root/trunk@6468 27541ba8-7e3a-0410-8455-c3a389f83636