m Philippe:
TChain::Add was broken since the latest improvement to the
regular expression.  Now it seems that:
    c = new TChain("T2");
    c->Add("event.root/T");
is taken has a regular expression.  This would have been 'okay' if it was
not for the fact that the regular expression par of TChain::Add never
supported the '/treename' syntax.  In addition, if (and only if) a path was
specified then the filename had to end with '.root' (if the path was not
specified the filename could end in anything the user wanted!).

This patch fix those problems (and add support for any file name length!)


git-svn-id: http://root.cern.ch/svn/root/trunk@7005 27541ba8-7e3a-0410-8455-c3a389f83636