from Philippe:
The current implementation of TROOT::GetGlobalFunction had the side effect of creating the byte code for ALL the cint
global function (even if they crash cint but are never used :)).
Also the current code was different for Windows and Unix.

This provides an implementation that is the same on all platform and does not provokes the cint compilation.

Also it now fully relies on TFunction::GetMangledName which is now cached.  The fact that it is not cached was also
causing problems when the file was reload.


git-svn-id: http://root.cern.ch/svn/root/trunk@7030 27541ba8-7e3a-0410-8455-c3a389f83636