Fix by Philippe for a problem reported by Gordon Watts when installing under VC++7.1.NET (was OK under 7.0)
I finally discovered the source of the problem (library is generated, compiled, linked and loaded properly but the
symbol are not available).

Because of one bug in rootcint and one bug in CINT and the fact that you are using a period ('.') in one the directory
name, the DLLID of all your libraries was the same (hence all but the first one thought they were already loaded into
the CINT dictionaries).


git-svn-id: http://root.cern.ch/svn/root/trunk@7039 27541ba8-7e3a-0410-8455-c3a389f83636