From Philippe:
This update to TSystem::CompileMacro adds a new option 'c' which tells ACLiC to compile and link the shared library but
do not load it.  This allow using ACLiC in makefile to link shared library that are not closed (except on platform that
do not support this).

This update also corrects the returns value of TSystem::CompileMacro (now always returns true in all cases of success
and false in case of failure) and corrects the indentation.


git-svn-id: http://root.cern.ch/svn/root/trunk@7084 27541ba8-7e3a-0410-8455-c3a389f83636