From Philippe:
This patch solves 2 problems.

If a simple tree was made the friend of a chain, the simple tree was loaded using the 'entry in the tree (inside the
chain)' instead of the 'entry in the chain'.  This resulted in ONLY the first part of the simple tree to be used (over
and over again)!

TTree::Scan (Query and CopyTree) could not handle the case where a TChain was made a friend of a TTree (and problably
more cases of uneven length chain friends).  This was
because, it never detected that a new file was open since it only looked a the main Tree (and none of its friend).  In
consequence the formula were not updated when they should.


git-svn-id: http://root.cern.ch/svn/root/trunk@7153 27541ba8-7e3a-0410-8455-c3a389f83636