From Olivier:
In TGaxis and TAxis the time offset in now managed without using TDatime.
Using TDatime was not a good idea because its starting date is 1995. The
new mechanisms uses now 1970 (like UNIX) as starting date. Note that the
functionality and examples given in the HowTo page and documentation
remains exactly the same.


git-svn-id: http://root.cern.ch/svn/root/trunk@7233 27541ba8-7e3a-0410-8455-c3a389f83636