From Axel Naumann:
THtml will now create the following warning when documention an abstract
class: "This is an abstract class, constructors will not be documented.
Look at the header to check for available constructors.", where "header"
is linked to the header html file of the class. This warning appears right
above the method list.


git-svn-id: http://root.cern.ch/svn/root/trunk@7428 27541ba8-7e3a-0410-8455-c3a389f83636