From Philippe:
TFormula::Analyze now check whether an expression is numerical __before__ checking whether it is a defined variable.
Without this patch doing something as simple as:
        mytree->Draw(".5");
        mytree->Draw("goodvar*.5");
could fails depending on the structure of the tree.  In particular, if a new 'weird'
branch is added to a tree, those statement might now lead to core dumps event if
the new branch is not being used.  (The core dumps should not be happening in the
first place ... but this is more stable :) ).

In TTreeFormula, we now properly skip the case of a TClonesArray inside a split TClonesArray when looking for data
members and methods (when the branch name is not fully explicit).


git-svn-id: http://root.cern.ch/svn/root/trunk@7453 27541ba8-7e3a-0410-8455-c3a389f83636