From Philippe:  this is a patch that implements the forwarding of kerberos credentials from the client to the server. In particular this allow the usage of kerberos both from the client to the proof server and from the proof master to the proof slave.
I also added the ability to specify a target user in rootauthrc when
selecting the kerberos authentification (i.e. master and slave will run
as this user is allowed).

In order to implement these 2 features, the client and server need to have
a few extra communications (send the target user name and the credentials),
to support this I had to increment the protocol numbers.


git-svn-id: http://root.cern.ch/svn/root/trunk@7471 27541ba8-7e3a-0410-8455-c3a389f83636