From Philippe:
This update makes any selector created by the TTreePlayer directly from a source file local to the current TTreePlayer
object (instead of being a static).

It also insures that TProofPlayer and TTreePlayer do not call delete for an object of
a class which library was unloaded (since the code is gone, this would lead to a core
dump).


git-svn-id: http://root.cern.ch/svn/root/trunk@7664 27541ba8-7e3a-0410-8455-c3a389f83636