From Damir and Olivier:
Correct a few remaining problems for time axis. Now, the time stored in
the format is always expressed in UTC. The local time is calculated
only at display time if the "gmt" option of SetTimeOffset is not used.
This allows to send a plot to someone in another time zone without
discrepancies on the time displayed.
If the plot uses "gmt", it will always show the same, whatever time
zone. If not, the time displayed will be the one local to the
displaying machine.


git-svn-id: http://root.cern.ch/svn/root/trunk@7668 27541ba8-7e3a-0410-8455-c3a389f83636