From Philippe
This patch makes sure the formula are properly refreshed when using a chain.

It removes the printing of some (spurious) empty lines when using a selection criterion.
I.e. Without a selection criterion, there is always a least one line per entry (if is empty) while with a selection
criterion, an entry is printed if and only if a least one of the sub-element exist and match the criterion.


git-svn-id: http://root.cern.ch/svn/root/trunk@7777 27541ba8-7e3a-0410-8455-c3a389f83636