From Philippe:
Using valgrind on TBench, I found some uninitialized memory in TLeaf
in the case of a TLeafElement belonging to a split TClonesArray.

The uninitialized memory came from the fact that the constructor was
not initializing everything if it could not find the leafcount.
In the case described above, the leafcount is not available until
after the constructor of the leaf (or so it seems) and is explicitly
set later using TLeaf::SetLeafCount.

This patch insures that more data members are initialized even in
case of seemingly incorrect configuration and insures that the Zombie
state is removed if the leafcount becomes available after the end of
the constructor.


git-svn-id: http://root.cern.ch/svn/root/trunk@7783 27541ba8-7e3a-0410-8455-c3a389f83636