Implement a new data type Double32_t (in Rtypes.h). In memory a Double32_t is like a Double_t (8 bytes). When written to a file, the type is converted to a Float 4 bytes. The newdata type is accepted as a simple variable, as a fixed length array or a variable length array, like:   Double32_t  fPt:   Double32_t  fVertex[3];   Int_t       fNtracks;   Double32_t *fPx;  //[fNtracks]
The new data type is supported in all I/O modes (TKey or TTree)

The automatic schema evolution algorithm accepts this new type.
For example a class in a file containing a Double_t data member
can be read by a new class where the member is now of type Double32_t
and vice-versa.


git-svn-id: http://root.cern.ch/svn/root/trunk@7804 27541ba8-7e3a-0410-8455-c3a389f83636