From Philippe:
icc8 defines some of the gcc symbols.  This can leads to a lot of conflicts in the cintdlls (including string).
This patches fixes the problem.

Extra for icc8's documentation

Source Compatibility with gcc 3.2
The predefined macros enabled by gcc are now enabled by the Intel Compiler. For example, the Intel Compiler now defines
the __GNUC__, __GNUC_MINOR__, and GNUC_PATCHLEVEL__ macros. If you do not want these macros to be defined, you can
specify the -no-gcc option.
Additionally, this version of the Intel C++ Compiler for Linux uses the C headers shipped with the version of Linux you
are running on, with the exception of two small substitute headers used only by the Itanium compiler for Itanium
applications. Also, the -cxxlib-gcc switch now compiles the Gnu C++ library headers if you use the -ansi language mode.


git-svn-id: http://root.cern.ch/svn/root/trunk@7893 27541ba8-7e3a-0410-8455-c3a389f83636