From Philippe: This patch repairs several issues dealing with loading a library and a corresponding root file when they define/contain stl containers.
TROOT.cxx:
   The dictionary loading has a recursive behavior which implies
that some entry are actually loaded several time (mostly harmless
but hard to prevent).  This leads to a case where the const char*
is modified during the loading of the dictionary.

TClass.h and TClasss.cxx:
   Add the following convenience static functions:
      TClass::GetClass  equivalent to gROOT->GetClass
      TClass::GetDict   equivalent to TClassTable::GetDict
   Fix a problem where the TClass object stored in the current
class TStreamerInfo were not refreshed when the TClass object
was replace.  The fix is to insure that the current TClass is
in the listOfClasses when the replacement might happen

TTree.cxx and TStreamerInfo.cxx
   If an stl container is a base class, it ends up with a
TStreamerInfo on the file.  However TStreamerInfo for stl
containers are not always built.   Anyway the StreamerInfo
is not used.  So it should not be on the file (fix in TTree).
However for backward compatibility we now handle the case in
BuildCheck.


git-svn-id: http://root.cern.ch/svn/root/trunk@8191 27541ba8-7e3a-0410-8455-c3a389f83636