From Philippe: In the previous implementation in the case where the user set the pointer (used by SetBranchAddress) pointing to a TClonesArray to zero after constructing the branch that is split, NOTHING was saved in the main branch (where the size of the collection is supposed to go).
In this new version, we insert the value zero that when reading back,
it is considered as an empty collection.

In addition we updated the read mechanism so that it detects case where
the value read for the index is greater than the maximum.
When this is the case, we check  whether we are in the case where
nothing has been saved in which case we set the length to 0 (instead of
the random value used in the previous implementation) and rewind the
buffer appropriately.
If we are not in this known case, we issue an error message and set the
length to 0 and we do NOT rewind the buffer.


git-svn-id: http://root.cern.ch/svn/root/trunk@8343 27541ba8-7e3a-0410-8455-c3a389f83636