From Gerri: Following a thread in roottalk I have found a bug causing authentication problems at master/slave level.
This is due to the fact that, as it is now, ClearAuth saves
the password-hash in fgPasswd, instead of the password. And
that is the quantity transmitted to the master for authentication
to the slave. This of course cannot work if the pass-hash
on the slave is different from the one on the master (different
salt).

I have fixed this by filling fgPasswd with orginal passwords.


git-svn-id: http://root.cern.ch/svn/root/trunk@8506 27541ba8-7e3a-0410-8455-c3a389f83636