New functions added to TBuffer in view of the coming XML interface    virtual void IncrementLevel(TStreamerInfo*)    virtual void SetStreamerElementNumber(Int_t)    virtual void DecrementLevel(TStreamerInfo*)
The IncrementLevel & Decrementlevel functions are called by the
TClass::ReadBuffer and TClassWriteBuffer functions.

The function SetStreamerElementNumber is called by
TStreamerInfo::ReadBuffer and TStreamerInfoWriteBuffer
to pass to the buffer the loop index.

This new version of TBuffer is still experimental.
If ROOT is compiled with the option R__BufferVirtual, several
functions from TBuffer are declared virtual. The default is
to have the TBuffer functions non virtual.

We are currently investigating the performance penalty induced
by making these functions virtual.


git-svn-id: http://root.cern.ch/svn/root/trunk@8812 27541ba8-7e3a-0410-8455-c3a389f83636