new methods:
void   TTimer::SetInterruptSyscall(Bool_t)
Bool_t TTimer::InterruptsSyscall()

When the argument is true the a-synchronous timer (SIGALRM) signal
handler is set so that interrupted syscalls will not be restarted
by the kernel. This is typically used in case one wants to put a
timeout on an I/O operation. By default interrupted syscalls will
be restarted.


git-svn-id: http://root.cern.ch/svn/root/trunk@8821 27541ba8-7e3a-0410-8455-c3a389f83636