From Axel Naumann:
TChain's and TTree's LoadTree behave differently when requesting an
entry which doesn't exist. A loop

TTree* tree=...
Int_t entry=0;
while (tree->LoadTree(entry)>=0) {
   tree->GetEntry(entry++);
   ...
}
works for tree being a TChain, but not for TTrees, as TTree::LoadTree
will return the entry number even if it's out of range. With the
attached patch TTree::LoadTree(Int_t) returns -2 if the entry does not
exist, just as TChain::LoadTree does.


git-svn-id: http://root.cern.ch/svn/root/trunk@8856 27541ba8-7e3a-0410-8455-c3a389f83636